requirements.txt
pyserial
pymavlink
loguru